requirements.txt
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
contourpy
cramjam==2.11.0
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
fonttools==4.59.0
frozendict==2.4.6
fsspec==2025.7.0
graphviz==0.21
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.10.5
matplotlib-inline==0.1.7
multitasking==0.0.12
nest-asyncio==1.6.0
numpy
packaging==25.0
pandas
parso==0.8.4
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
#pywin32==311
pyzmq==27.0.1
requests==2.32.4
yfinance
scikit-learn
pyarrow
fastparquet
scipy
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
